# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Papi Foundation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /community/members:
    get:
      operationId: listCommunityMembers
      tags: [community]
      summary: List community members
      responses:
        "200":
          description: Community members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommunityMember"
    post:
      operationId: joinCommunity
      tags: [community]
      summary: Join the community
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityMemberInput"
      responses:
        "201":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityMember"
...
  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List blog posts
      responses:
        "200":
          description: Blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/{id}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
...
components:
  schemas:
...
    CommunityMember:
      type: object
      required: [id, name, email, country, memberType, joinedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        country:
          type: string
        memberType:
          type: string
          enum: [individual, organization, volunteer]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        joinedAt:
          type: string
    CommunityMemberInput:
      type: object
      required: [name, email, country, memberType]
      properties:
        name:
          type: string
        email:
          type: string
        country:
          type: string
        memberType:
          type: string
          enum: [individual, organization, volunteer]
        bio:
          type: string
        avatarUrl:
          type: string
...
    BlogPost:
      type: object
      required: [id, title, excerpt, content, author, category, imageUrl, publishedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        author:
          type: string
        category:
          type: string
          enum: [story, report, press, impact]
        imageUrl:
          type: string
        publishedAt:
          type: string
    BlogPostInput:
      type: object
      required: [title, excerpt, content, author, category, imageUrl]
      properties:
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        author:
          type: string
        category:
          type: string
          enum: [story, report, press, impact]
        imageUrl:
          type: string
...